fix: refactoring lang page
